move from io.kixi.* to io.kixi.core.*
