--- src/main/kotlin/io/kixi/Duration+.kt
@@ -1,7 +1,0 @@
-package io.kixi
-
-import java.time.Duration
-
-fun Duration.kiFormat(zeroPad:Boolean = false): String {
-    return Ki.formatDuration(this, zeroPad)
-}
