--- src/main/kotlin/io/kixi/KiTZ.kt
@@ -1,119 +1,0 @@
-package io.kixi
-
-import java.time.ZoneOffset
-
-/**
- * A Ki time zone (KiTZ). See
- * <a href="https://github.com/kixi-io/Ki.Docs/wiki/Ki-Time-Zone-Specification">Ki Time Zone Specification</a>
- */
-class KiTZ {
-    companion object {
-        @JvmStatic val offsets = mapOf<String, ZoneOffset>(
-            "AR/ART"	to ZoneOffset.of("-3"),
-            "AU/ACST"	to ZoneOffset.of("+09:30"),
-            "AU/ACDT"	to ZoneOffset.of("+10:30"),
-            "AU/AEST"	to ZoneOffset.of("+10"),
-            "AU/AEDT"	to ZoneOffset.of("+11"),
-            "AU/AWST"	to ZoneOffset.of("+8"),
-            "AU/NFT"	to ZoneOffset.of("+11"),
-            "AU/NFDT"	to ZoneOffset.of("+12"),
-            "BR/ACT"	to ZoneOffset.of("-5"),
-            "BR/AMT"	to ZoneOffset.of("-4"),
-            "BR/BRT"	to ZoneOffset.of("-3"),
-            "BR/FNT"	to ZoneOffset.of("-2"),
-            "CA/AST"	to ZoneOffset.of("-4"),
-            "CA/ADT"	to ZoneOffset.of("-3"),
-            "CA/CST"	to ZoneOffset.of("-6"),
-            "CA/CDT"	to ZoneOffset.of("-5"),
-            "CA/EST"	to ZoneOffset.of("-5"),
-            "CA/EDT"	to ZoneOffset.of("-4"),
-            "CA/MST"	to ZoneOffset.of("-7"),
-            "CA/MDT"	to ZoneOffset.of("-6"),
-            "CA/NST"	to ZoneOffset.of("-03:30"),
-            "CA/NDT"	to ZoneOffset.of("-02:30"),
-            "CA/PST"	to ZoneOffset.of("-8"),
-            "CA/PDT"	to ZoneOffset.of("-7"),
-            "CN/CST"	to ZoneOffset.of("+8"),
-            "CO/COT"	to ZoneOffset.of("-5"),
-            "CZ/CET"	to ZoneOffset.of("+1"),
-            "CZ/CEST"	to ZoneOffset.of("+2"),
-            "EG/EET"	to ZoneOffset.of("+2"),
-            "ET/EAT"	to ZoneOffset.of("+3"),
-            "FR/CET"	to ZoneOffset.of("+1"),
-            "FR/CEST"	to ZoneOffset.of("+2"),
-            "DE/CET"	to ZoneOffset.of("+1"),
-            "DE/CEST"	to ZoneOffset.of("+2"),
-            "GR/EET"	to ZoneOffset.of("+2"),
-            "GR/EEST"	to ZoneOffset.of("+3"),
-            "IN/IST"	to ZoneOffset.of("+05:30"),
-            "ID/WIB"	to ZoneOffset.of("+7"),
-            "ID/WITA"	to ZoneOffset.of("+8"),
-            "ID/WIT"	to ZoneOffset.of("+9"),
-            "IR/IRST"	to ZoneOffset.of("+03:30"),
-            "IR/IRDT"	to ZoneOffset.of("+04:30"),
-            "IQ/AST"	to ZoneOffset.of("+3"),
-            "IE/IST"	to ZoneOffset.of("+1"),
-            "IT/CET"	to ZoneOffset.of("+1"),
-            "IT/CEST"	to ZoneOffset.of("+2"),
-            "JP/JST"	to ZoneOffset.of("+9"),
-            "KE/EAT"	to ZoneOffset.of("+3"),
-            "MX/CST"	to ZoneOffset.of("-6"),
-            "MX/CDT"	to ZoneOffset.of("-5"),
-            "MX/EST"	to ZoneOffset.of("-5"),
-            "MX/MST"	to ZoneOffset.of("-7"),
-            "MX/MDT"	to ZoneOffset.of("-6"),
-            "MX/PST"	to ZoneOffset.of("-8"),
-            "MX/PDT"	to ZoneOffset.of("-7"),
-            "NL/CET"	to ZoneOffset.of("+1"),
-            "NL/CEST"	to ZoneOffset.of("+2"),
-            "NL/AST"	to ZoneOffset.of("-4"),
-            "NZ/NZST"   to ZoneOffset.of("+12"),
-            "NZ/NZDT"   to ZoneOffset.of("+13"),
-            "NG/WAT"	to ZoneOffset.of("+1"),
-            "PK/PST"	to ZoneOffset.of("+5"),
-            "PL/CET"	to ZoneOffset.of("+1"),
-            "PL/CEST"	to ZoneOffset.of("+2"),
-            "RO/EET"	to ZoneOffset.of("+2"),
-            "RO/EEST"	to ZoneOffset.of("+3"),
-            "RU/KALT"	to ZoneOffset.of("+2"),
-            "RU/MSK"	to ZoneOffset.of("+3"),
-            "RU/SAMT"	to ZoneOffset.of("+4"),
-            "RU/YEKT"	to ZoneOffset.of("+5"),
-            "RU/OMST"	to ZoneOffset.of("+6"),
-            "RU/KRAT"	to ZoneOffset.of("+7"),
-            "RU/IRKT"	to ZoneOffset.of("+8"),
-            "RU/YAKT"	to ZoneOffset.of("+9"),
-            "RU/VLAT"	to ZoneOffset.of("+10"),
-            "RU/MAGT"	to ZoneOffset.of("+11"),
-            "RU/PETT"	to ZoneOffset.of("+12"),
-            "SA/AST"	to ZoneOffset.of("+3"),
-            "SG/SST"    to ZoneOffset.of("+8"),
-            "ZA/SAST"	to ZoneOffset.of("+2"),
-            "KR/KST"	to ZoneOffset.of("+9"),
-            "ES/CET"	to ZoneOffset.of("+1"),
-            "ES/CEST"	to ZoneOffset.of("+2"),
-            "SE/CET"	to ZoneOffset.of("+1"),
-            "SE/CEST"	to ZoneOffset.of("+2"),
-            "CH/CET"	to ZoneOffset.of("+1"),
-            "CH/CEST"	to ZoneOffset.of("+2"),
-            "TR/FET"	to ZoneOffset.of("+3"),
-            "UA/EET"	to ZoneOffset.of("+2"),
-            "UA/EEST"	to ZoneOffset.of("+3"),
-            "GB/DST"	to ZoneOffset.of("+1"),
-            "US/AST"	to ZoneOffset.of("-4"),
-            "US/AKST"	to ZoneOffset.of("-9"),
-            "US/AKDT"	to ZoneOffset.of("-8"),
-            "US/CHST"	to ZoneOffset.of("+10"),
-            "US/CST"	to ZoneOffset.of("-6"),
-            "US/CDT"	to ZoneOffset.of("-5"),
-            "US/EST"	to ZoneOffset.of("-5"),
-            "US/EDT"	to ZoneOffset.of("-4"),
-            "US/HST"	to ZoneOffset.of("-10"),
-            "US/MST"	to ZoneOffset.of("-7"),
-            "US/MDT"	to ZoneOffset.of("-6"),
-            "US/PST"	to ZoneOffset.of("-8"),
-            "US/PDT"	to ZoneOffset.of("-7"),
-            "US/SST"	to ZoneOffset.of("-11")
-        )
-    }
-}
