+++ src/main/kotlin/io/kixi/core/BigDecimal+.kt
@@ -1,0 +1,10 @@
+package io.kixi.core
+
+import java.math.BigDecimal
+
+/**
+ * Does this represent a whole number?
+ */
+val BigDecimal.whole: Boolean
+    get() =
+        this.signum() == 0 || this.scale() <= 0 || this.stripTrailingZeros().scale() <= 0
