+++ src/main/kotlin/io/kixi/core/Duration+.kt
@@ -1,0 +1,7 @@
+package io.kixi.core
+
+import java.time.Duration
+
+fun Duration.kiFormat(zeroPad:Boolean = false): String {
+    return Ki.formatDuration(this, zeroPad)
+}
