+++ src/main/kotlin/io/kixi/core/uom/IncompatibleUnitsException.kt
@@ -1,0 +1,6 @@
+package io.kixi.core.uom
+
+import java.lang.RuntimeException
+
+class IncompatibleUnitsException(from:Unit, to:Unit) :
+    RuntimeException("Can't convert from ${from::class.java.simpleName} to ${to::class.java.simpleName}")
