+++ src/main/kotlin/io/kixi/core/uom/NoSuchUnitException.kt
@@ -1,0 +1,6 @@
+package io.kixi.core.uom
+
+import java.lang.RuntimeException
+
+class NoSuchUnitException(symbol:String) :
+    RuntimeException("Unit for symbol $symbol is not recognized.")
