--- src/main/kotlin/io/kixi/uom/IncompatibleUnitsException.kt
@@ -1,6 +1,0 @@
-package io.kixi.uom
-
-import java.lang.RuntimeException
-
-class IncompatibleUnitsException(from:Unit, to:Unit) :
-    RuntimeException("Can't convert from ${from::class.java.simpleName} to ${to::class.java.simpleName}")
