--- src/main/kotlin/io/kixi/uom/NoSuchUnitException.kt
@@ -1,6 +1,0 @@
-package io.kixi.uom
-
-import java.lang.RuntimeException
-
-class NoSuchUnitException(symbol:String) :
-    RuntimeException("Unit for symbol $symbol is not recognized.")
